ce_videnc: add setting to define image height alignment.
Some DM36x CE codecs has limitations on the image height supported. With
this setting we allow codec sub-class to define the corresponding height
limitation. So the class can fail during configuration if the given height
alignment is not fulfilled. On the other hand, some codecs even if the
height alignment is not correctly a padding is allowed at the end of the
buffer so the codec can process it, with this new setting the base class
will check it the given buffer has the required padding.